requirements.txt
pandas
fastapi
uvicorn
pyarrow
openpyxl
psycopg2-binary
python-dotenv